requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Supabase
supabase>=2.27.0
postgrest==0.13.0
websockets>=11,<16

# OpenAI
openai>=1.54.0

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
